requirements.txt
django==3.2.9
djangorestframework==3.12.4
python-dotenv==0.19.1
psycopg2==2.9.1
django-mptt==0.13.4
djangorestframework-recursive==0.1.2